+++ stress/v2/statement/set.go
@@ -4,8 +4,7 @@ import (
 	"fmt"
 	"strings"
 
-	influx "github.com/influxdata/influxdb/client/v2"
-	"github.com/influxdata/influxdb/stress/v2/ponyExpress"
+	"github.com/influxdata/influxdb/stress/v2/stress_client"
 )
 
 // SetStatement set state variables for the test
@@ -15,7 +14,7 @@ type SetStatement struct {
 
 	StatementID string
 
-	Tracer *ponyExpress.Tracer
+	Tracer *stressClient.Tracer
 }
 
 // SetID statisfies the Statement Interface
@@ -24,57 +23,30 @@ func (i *SetStatement) SetID(s string) {
 }
 
 // Run statisfies the Statement Interface
-func (i *SetStatement) Run(s *ponyExpress.StoreFront) {
-
-	// Set the Tracer
-	i.Tracer = ponyExpress.NewTracer(make(map[string]string))
-
-	// Create a new Directive
-	d := ponyExpress.NewDirective(strings.ToLower(i.Var), strings.ToLower(i.Value), i.Tracer)
-
+func (i *SetStatement) Run(s *stressClient.StressTest) {
+	i.Tracer = stressClient.NewTracer(make(map[string]string))
+	d := stressClient.NewDirective(strings.ToLower(i.Var), strings.ToLower(i.Value), i.Tracer)
 	switch d.Property {
-
-	// Needs to be set on both StoreFront and ponyExpress
+	// Needs to be set on both StressTest and stressClient
 	// Set the write percison for points generated
 	case "precision":
 		s.Precision = d.Value
-
-		// Increment the tracer
 		i.Tracer.Add(1)
 		s.SendDirective(d)
-
-	// Lives on StoreFront
+	// Lives on StressTest
 	// Set the date for the first point entered into the database
 	case "startdate":
 		s.Lock()
 		s.StartDate = d.Value
 		s.Unlock()
-
-	// Lives on StoreFront
+	// Lives on StressTest
 	// Set the BatchSize for writes
 	case "batchsize":
 		s.Lock()
 		s.BatchSize = parseInt(d.Value)
 		s.Unlock()
-
-	// Lives on StoreFront
-	// Reset the ResultsClient to have a new address
-	case "resultsaddress":
-		s.Lock()
-		s.SetResultsClient(influx.HTTPConfig{Addr: fmt.Sprintf("http://%v/", d.Value)})
-		s.Unlock()
-
-	// TODO: Make TestName actually change the reporting DB
-	// Lives on StoreFront
-	// Set the TestName that controls reporting DB
-	case "testname":
-		s.Lock()
-		s.TestName = d.Value
-		s.Unlock()
-
-	// All other variables live on ponyExpress
+	// All other variables live on stressClient
 	default:
-		// Increment the tracer
 		i.Tracer.Add(1)
 		s.SendDirective(d)
 	}
@@ -82,6 +54,6 @@ func (i *SetStatement) Run(s *ponyExpress.StoreFront) {
 }
 
 // Report statisfies the Statement Interface
-func (i *SetStatement) Report(s *ponyExpress.StoreFront) string {
+func (i *SetStatement) Report(s *stressClient.StressTest) string {
 	return fmt.Sprintf("SET %v = '%v'", i.Var, i.Value)
 }