+++ stress/v2/stress_client/tracer.go
@@ -1,4 +1,4 @@
-package ponyExpress
+package stressClient
 
 import (
 	"sync"